+++ docs/src/reference/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## `v1.4.2`
 
 - **added** an `openstreetmap` method (not exported) to get various GeoJSON limits from plan-text queries
+- **added** a `simplify` and `simplify!` method to remove some complexity from OSM polygons
 - **fixed** the vignettes so they don't use GADM
 
 ## `v1.4.1`